+++ CotEditor/Sources/Settings Window/Syntax Editor/SyntaxCompletionEditView.swift
@@ -64,6 +64,10 @@ struct SyntaxCompletionEditView: View {
                 AddRemoveButton($items, selection: $selection, newItem: Item()) { item in
                     self.focusedField = item.id
                 }
+                Text("\(self.items.count) items", tableName: "SyntaxEditor")
+                    .monospacedDigit()
+                    .foregroundStyle(.secondary)
+                    .controlSize(.small)
                 Spacer()
                 HelpLink(anchor: "syntax_highlight_settings")
             }