+++ CotEditor/Sources/Settings Window/Syntax Editor/SyntaxHighlightEditView.swift
@@ -111,6 +111,10 @@ struct SyntaxHighlightEditView: View {
                 AddRemoveButton($items, selection: $selection, newItem: Item()) { item in
                     self.focusedField = item.id
                 }
+                Text("\(self.items.count) items", tableName: "SyntaxEditor")
+                    .monospacedDigit()
+                    .foregroundStyle(.secondary)
+                    .controlSize(.small)
                 Spacer()
                 HelpLink(anchor: self.helpAnchor)
             }